+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - File size displayed
 - LICENSE: AGPL v3
 - Option (on by default) to use the category in tags for file uploads.
+- Option to only automatically select a category, persona or workflow once.
 
 ### Changed